requirements.txt
requests
pytest==5.3.5
selenium==4.0.0a3
pytest-html==2.0.1
webdriver-manager==2.3.0
allure-pytest>=2.8.11
pytest-rerunfailures>=8.0
pytest-testconfig>=0.2.0